requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
pykrx>=1.0.0
